beginning work on shader graph
static render system is being put on hold for a second.
I need to make shaders for the static rendering system, with countless variations for each implementation.
I believe a Shader Graph system will make it easier, or at least lead me to learn the necessary concepts, to create an easier way to generate the variants, without using the convoluted system found in ShaderBuilder.h.

The system in found in ShaderBuilder.h largely arose from the need to include different material sets, and was extended to account for bones and instancing. It's a mess. Even if I wanted to look at it now, I'd have to stare at it for a solid minute (hours) to figure out what it does, and still trace the steps to extend functionality. Not a good system.
I'd like to replace ALL hand written shaders with graphs, but I'll have to see where I end up at the end of this.

Why am i writing such a long description
