+++ examples/RockSample/src/GUI/MainMenuMM.cpp
@@ -14,12 +14,14 @@ namespace EWE {
 		clickText.emplace_back("Audio Settings", 0.f, 680.f * heightRescaling, TA_left, 3.f, screenWidth, screenHeight);
 
 		clickText.emplace_back("Exit", 0.f, 860.f * heightRescaling, TA_left, 3.f, screenWidth, screenHeight);
+		clickText.emplace_back("Shader Generation", 0.f, 520.f * heightRescaling, TA_left, 3.f, screenWidth, screenHeight);
 		//clickText.emplace_back("Level Builder", 0.f, 760.f * heightRescaling, TA_left, 3.f, screenWidth, screenHeight);
 		
 	}
 
 	void MainMenuMM::processClick(double xpos, double ypos) {
-		std::pair<UIComponentTypes, int8_t> returnValues = MenuModule::checkClick(xpos, ypos);
+		std::pair<UIComponentTypes, int16_t> returnValues = MenuModule::checkClick(xpos, ypos);
+
 
 		if (returnValues.first == UIT_ClickTextBox) {
 			//printf("object comparison : %lu:%lu \n", (uint64_t)this, (uint64_t)uiPointer);
@@ -38,10 +40,12 @@ namespace EWE {
 			}
 			case 2: {
 				clickReturns.push(MCR_ExitProgram);
+				break;
 			}
 			case 3: {
 				printf("should not be able to enter LB mode, probably a crash \n");
-				clickReturns.push(MCR_switchToLB);
+				clickReturns.push(MCR_swapToShaderGen);
+				break;
 			}
 			
 			default: {
