+++ include/EWEngine/ShaderGraph/Edge.h
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace EWE {
+	//
+	struct Edge {
+
+	};
+}
+
