+++ include/EWEngine/ShaderGraph/Node.h
@@ -1,0 +1,14 @@
+#pragma once
+
+namespace EWE {
+	//
+	struct Edge;
+
+	struct Node {
+		Edge* inEdge{nullptr};
+		Edge* outEdge{nullptr};
+
+		void execute() {}
+	};
+}
+
