+++ src/UIHandler.cpp
@@ -73,7 +73,7 @@ namespace EWE {
 #endif
 
 
-	void UIHandler::drawMenuMain(VkCommandBuffer commandBuffer, bool displayingRenderInfo) {
+	void UIHandler::drawOverlayText(VkCommandBuffer commandBuffer, bool displayingRenderInfo) {
 
 		if (overlay) {
 			overlay->drawText();
